refactor(test-base): Update base testing classes
