+++ app/tests/test_main.py
@@ -1,4 +1,4 @@
-from app.settings import settings
+from app.core.config import settings
 from app.tests.test_base import TestBase
 
 
@@ -7,8 +7,9 @@ class TestMain(TestBase):
         response = self.client.get("/")
         assert response.status_code == 200
         assert response.json() == {
-            "app_name": settings.app_name,
-            "app_description": settings.app_description,
-            "app_version": settings.app_version,
+            "app_name": settings.APP_NAME,
+            "app_description": settings.APP_DESCRIPTION,
+            "app_version": settings.APP_VERSION,
             "docs_url": "http://testserver/docs",
+            "openapi_url": "http://testserver/openapi.json",
         }
